requirements.txt
# Core Framework
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-community>=0.3.0

# Vector Database
chromadb>=0.4.0

# Embeddings & NLP
sentence-transformers>=2.2.0
transformers>=4.35.0
torch>=2.0.0
spacy>=3.7.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0

# API Framework
fastapi>=0.109.0
uvicorn>=0.25.0
websockets>=12.0
python-multipart>=0.0.6

# Data Ingestion
feedparser>=6.0.0
httpx>=0.26.0
beautifulsoup4>=4.12.0
aiohttp>=3.9.0

# Scheduling
apscheduler>=3.10.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Sentiment Analysis (FinBERT)
transformers>=4.36.0
torch>=2.0.0
accelerate>=0.25.0

# Production Caching
redis>=5.0.0

# Monitoring & Observability  
prometheus-client>=0.19.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
isort>=5.12.0
